refactor: update tests to use test co-location
